Backend

+++ backend/estagio-poc/src/main/java/br/ufpr/estagio/modulo/repository/ApoliceRepository.java
@@ -1,11 +1,9 @@
 package br.ufpr.estagio.modulo.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.*;
 
 import br.ufpr.estagio.modulo.model.Apolice;
 
-public interface ApoliceRepository extends JpaRepository<Apolice, Integer>{
+public interface ApoliceRepository extends JpaRepository<Apolice, Long>{
 
 }
